Describes how to run spark master and submit jobs in standalone mode
Now, the start overhead of spark-shell is eliminated and we completely rely on a master to which jobs are submitted each time. spark-shell introduced a lot of overhead since it starts a master process and a scala interpreter shell on it each time a program is executed. The program was simply directed via stdin to the interpreter shell line by line which explains the significant overhead when executing spark programs in this fashion.